move tests not required by the PDF into folder called extra
This it to prevent confusion on what is mandatory.
A lot of people are running the tester and the vague PDF
combined with tests not being well organized, creates pure chaos.

We should aim to improve the overall minishell experience, by
removing as much vagueness as possible from the project!